Add files via upload
Not sure what I'm doing here - I was following the prompts to get access to (or a copy of) Visual BASIC.

I've been using BASIC for over 50 years and I feel comfortable with it 
I don't look for speed - I look for code that I can understand when I look at it the day after I wrote it.

Anyway - I'm trying to get the web page RAILWAY.HTM working as I visualise it.  

It is written in bog-standard HTML using Notepad - an environment I am also comfortable with after 30 years of practice.

The problem is that pressing a button which links using <a href . . .  operates functions quite happily, but if I operate the same code  directly, it doesn't.
I'm hoping that VB will provide an answer - I used VB6 quite happily to automatically 'scrape' databases from the wwweb in its early days.

The system is based on four NODEMCU boards (Tom, Dick, Harry are plain and LUCY has an OLED screen.
They each set up as servers with a local MDNS, so that, say, http://Dick/1 is understood when entered to a browser using my router.
It corresponds to the 'Local' button.  When Dick gets the message, he operates a servo, and redirects to Lucy, who notes the change on her screen quite happily.
The Arduino code is simple . . . 
 server.on("/1",   handleOne );  near the top of the code 
and
void handleOne() {  current_route = "Local" ;  tellLucy(); }
further down the code
and it works admirably, however . . .
in the body of the code there are some timing loops which should set and reset the servos on a schedule
and I would like to use, say, 
if (millis() - last_millis > 30000){   handleOne();  tellLucy();  last_millis = millis(); }
but it refuses to behave.


Any advice would be welcome,